templates/autogen-agent-template/requirements.txt
pyautogen>=0.2.0
nerq>=1.0.0
python-dotenv>=1.0.0
